refactor: extract useUniverseGallery hook from UniverseBuilder page (#2532)
