+++ .changelog/NEXT.md
@@ -51,4 +51,5 @@
 - **[issue-2525] Extract shared `migrationMarker` helpers using `atomicWrite`.** Eleven boot-time migration/repair/reconcile scripts under `server/scripts/` each hand-rolled `markerExists`/`readMarker`/`writeMarker` on top of `readFile(...).catch(() => null)` and a non-atomic `writeFile(JSON.stringify(...))`, so a crash mid-write could leave a truncated marker a later boot misreads as "not applied." The three helpers now live once in `server/lib/migrationMarker.js` (anchored under `PATHS.data`, built on the existing `tryReadFile`/`atomicWrite`/`safeJSONParse`), are barrel- and README-cataloged, and carry their own unit test. Marker filenames, gate semantics, and payloads are unchanged; the write path is now atomic (temp file + rename).
 - **[issue-2522] Deep-link Tribe tabs and fix form label associations.** The Tribe page's active section (Circle / Map / Care Queue / Focus) lived in local `useState`, so a section couldn't be shared, bookmarked, or survive a reload. It now reads from a `?tab=` URL search param (default `circle` omitted for a clean `/tribe`) with a valid-tab fallback for stale/unknown values — matching the PromptManager convention. Separately, the `StylePresetPicker` style-preset `<label>` and the Loops "Interval" `<label>` had no `htmlFor`/`id` pairing; both now associate with their control via a `useId()`-generated id so click-to-focus and screen readers work.
 - **[issue-2527] Route client localStorage and keyboard activation through the shared `safeStorage`/`a11yKeyboard` helpers.** Several client sites bypassed the guarded `safeStorage` helpers with raw `window.localStorage` + hand-rolled `try/catch` (or `typeof window` guards), and several non-`<button>` clickables reimplemented the Enter/Space `onKeyDown` handler inline. Storage reads/writes in `Security.jsx`, `PipelineManuscriptEditor.jsx`, `voiceVisibility.js`, `MortalLoomBanner.jsx`, `WhileAwayWidget.jsx`, and `ManuscriptReadAloud.jsx` now go through `safeReadStorage`/`safeReadJsonStorage`/`safeWriteStorage`/`safeRemoveStorage`; keyboard activation in `IngredientPicker.jsx`, `ManuscriptReadAloud.jsx`, `EditorialHealthPanel.jsx`, `TribeCircleMap.jsx`, and `GenomeTab.jsx` now spreads `clickableProps(handler)` (which supplies `role`/`tabIndex`/`onKeyDown` via the shared `onActivateKeyDown`). Behavior is preserved — same storage keys, same JSON-parse fallbacks and corrupt-entry cleanup, same Enter/Space activation.
+- **[issue-2532] Extract the Universe Builder's page-level lightbox into a `useUniverseGallery` hook.** `UniverseBuilder.jsx` inlined ~160 lines of gallery/preview plumbing — the filename→image-sidecar metadata map, the `listImageGallery()` refetch effect (keyed on `runs.length` + a per-job-completion bump), the `previewItems` derivation across variations/composite sheets/style probes/canon+character-sheet refs (namespaced-key dedupe), and the `usePreviewRoute`/`useMediaPreviewActions`/`useMediaAnnotations` wiring. That cohesive concern now lives in `client/src/hooks/useUniverseGallery.js` (takes `{ draft, runsLength }`, returns `{ previewItems, preview, setPreview, previewActions, openPreviewByFilename, openVariationPreview, annotations, updateAnnotation, bumpGalleryRefresh }`), thinning the page shell toward the WritersRoom pattern. Pure structural refactor — behavior identical, no route-contract change; the UniverseBuilder page suite and hooks barrel test stay green. Further hook/panel extractions (draft/expand/canon/bulk-render for UniverseBuilder, `useShellSession` for Shell) remain.
 - **[issue-2545] Add `aria-label`s to icon-only controls and fix Templates/MessageDetail form labels.** Several icon-only buttons exposed their purpose only via `title` (unreliable for screen readers): the MessageDetail toolbar (Reply, voice-mode toggle, AI Reply, Refresh, Archive, Delete), MusicVideo scene move-up/move-down/delete, MediaCard hide/delete, the MediaJobsQueue per-job cancel, Review's edit save/cancel, and DuplicateGroup rename now carry an `aria-label` matching their `title`. The MessageDetail reply `<textarea>` (previously placeholder-only) gains an `aria-label`, and the Templates "Target Directory" `<label>` is now wired to its input via `htmlFor`/`id` so click-to-focus and screen readers work. Accessibility attributes only — no behavior or styling change.
